- updating import DateHelper - changing SkuPickup double to Float (skuPickupCbmAmount attribute)

+++ indianapolis-model/src/main/java/com/gdn/email/SkuPickup.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class SkuPickup {
     private String sku;
     private int skuPickupQuantity;
-    private double skuPickupCbmAmount;
+    private Float skuPickupCbmAmount;
 }
